requirements.txt
# === Core dependencies ===
requests
python-dotenv
textblob

# === Data analysis and visualization ===
pandas
matplotlib

# === Testing ===
pytest
pytest-mock
